add some string and move some class
