+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/match/MatchLeagueAdapter.kt
@@ -1,0 +1,27 @@
+package com.dicoding.picodiploma.sejiwaproject.ui.match
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.dicoding.picodiploma.sejiwaproject.R
+
+class MatchLeagueAdapter : RecyclerView.Adapter<MatchLeagueAdapter.MatchViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MatchViewHolder {
+        val view: View = LayoutInflater.from(parent.context).inflate(R.layout.item_match, parent, false)
+        return MatchViewHolder(view)
+    }
+
+    override fun getItemCount(): Int {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun onBindViewHolder(holder: MatchViewHolder, position: Int) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    inner class MatchViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+
+    }
+
+}
